Fix Inconsistent Interaction Events
Fix ordering of onInteractionStarted and onInteractionFinished events
in the AwesomeMap so that such events are not interleaved (ie: 2
onInteractionSTarted events followed by 2 onInteractionFinished
events). This could happen when interrupting a transformation via a
touch event.

Also fix the PeekInterceptor so that it played nice with this event
ordering by: not cancelling touch events, and running the scroll to
animation on the next tick.