Bounce At Boundaries on Swipe
When swiping near an AwesomeMap's boundaries, allow the content to
bounce into place to visually convey the boundary condition to the
user. Make this effect available via the BoundaryInterceptor's
'slow' mode.

Update PeekInterceptor to independently manage drags into
boundaries so that the BoundaryInterceptor can use 'slow' mode. This
has the beneficial side-effect of allowing users to drag the first and
last items in the list away from list boundaries and have the content
snap back upon release, thus visually conveying the boundary condition
in that case as well.