Account for top offset in hit test
Previously, HitTester.testItemMap() would not take into account
the offsetTop when testing for an event in the hit area. This is
used in peek mode to determine if there is a mouse move or click in a
particular bounding box on the page. The bounding box coordinates are
expected to be determined relative to the page, and the offsetTop
should be added to determine where this is.