Fix ScrollList ItemMap Hit Testing
Hit testing fails for item maps in a scroll list due to the calling
code passing in the current transform state for the list map, not for
the current item map. Fix by making the hit testing functions
responsible for looking up transform state and when testing item maps
look up the transform state for the item map, not the list map.