Fix Item Position Reported When Not In `flow` Mode
If the ScrollList is disabled during an interaction (between touch and
release events), the event positions reported are no longer relative
to the origin of its viewport, but are relative to the page origin
in the browser window. Currently the AwesomeMap will automatically
handle this in all cases.

For ScrollLists with nested AwesomeMaps for items, this correction
breaks down. The original event is handled first by the AwesomeMap for
the list itself, has its position normalized, and then is propogated
to the current item map, where a second normalization occurs. This
second normalization throws off the event position.