I believe this gets all of the uicomponent references
of scrollTo changed to scrollToItem.  KeyNavigator.js had some trailing
whitespace that my plugin wiped out.

At this point, I think this is done, but I need to ask Tim about how to
add unit tests and make sure that other repos are not using scrollTo.
(Renderer.js in wf-is-viewer is the only other place I have found so
far.)