Fix Sign of ScrollPosition Inside VerticalLayout
VerticalLayout confused the sign of the scroll position it tracks.
Translations coming out of the AwesomeMap for the list are sent into
the VerticalLayout#setScrollPosition method with the signs unchanged.
This is semantically incorrect, as translations use negative signs
to reflect positive scroll positions (scrolling to top position 100
is reflected by a translation along the y axis of -100).