Fixes placeholders occasionally not rendering
Swiping and causing many pages to scroll, and then moving the mouse
over the content while the content is scrolling would prevent
the placeholders for the new pages from rendering.

This occurred since mousemove events would be added to the
TransfomrationQueue after the dragend event. When the dragend event
was processed, the queue would not be empty, and so it would wait
to call ScrollList.render. When the mousemove events process they
do not trigger a render. So we'd end in a state where the
TransformationQueue was drained, but the placeholders in view would
not be rendered.

The solution is to ignore mousemove events while the Awesomemap is
being transformed. There should be no need to handle them at this
time anyway.