Fix Interrupted Simulated Swipe
When swiping on a trackpad, mouse wheel events are generated by the
hardware, and target the original source of the last "real" wheel
event. The virtualization of the ScrollList removes items as these
wheel events are processed, creating a situation where the scrolling
UX suddenly stops, as though a wall is hit.

Defer the removal of DOM elements until mouse wheel scrolling stops.
For consistency sake, do the same when handling swipe events.
Once interaction stops, the stale placeholders will be removed.